# job-api-project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portail Emploi RESTful",
        "description": "API pour la recherche et la gestion d'offres d'emploi, avec gestion des rôles (Admin, Employeur, Utilisateur).",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur API Principal"
        }
    ],
    "paths": {
...
        "/api/applications/job-offers/{jobOffer}": {
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Postuler à une offre d'emploi (Rôle : user requis)",
                "operationId": "applyToJob",
                "parameters": [
                    {
                        "name": "jobOffer",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cover_letter": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature soumise avec succès."
                    },
                    "403": {
                        "description": "Accès refusé."
                    },
                    "409": {
                        "description": "Déjà postulé."
                    },
                    "400": {
                        "description": "Offre inactive."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/mine": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Affiche les candidatures de l'utilisateur connecté (Rôle : user ou admin requis)",
                "operationId": "getUserApplications",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{jobOffer}/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Affiche les candidatures pour une offre donnée (Rôle : employeur ou admin requis)",
                "operationId": "getJobOfferApplications",
                "parameters": [
                    {
                        "name": "jobOffer",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidatures pour l'offre."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/{application}/status": {
            "put": {
                "tags": [
                    "Applications"
                ],
                "summary": "Met à jour le statut d'une candidature (Rôle : employeur ou admin requis)",
                "operationId": "updateApplicationStatus",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "reviewed",
                                            "accepted",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour avec succès."
                    },
                    "403": {
                        "description": "Accès refusé."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Personal Access Token",
                "scheme": "bearer"
            }
        }
    },
...
}